+++ app/src/main/kotlin/io/redgreen/timelapse/main/WorkbenchMenuBuilder.kt
@@ -1,10 +1,13 @@
 package io.redgreen.timelapse.main
 
+import io.redgreen.timelapse.foo.closeWindow
+import io.redgreen.timelapse.openrepo.view.OpenRepoScene
 import javafx.scene.Scene
 import javafx.scene.control.Menu
 import javafx.scene.control.MenuBar
 import javafx.scene.control.MenuItem
 import javafx.scene.layout.BorderPane
+import javafx.stage.Stage
 
 class WorkbenchMenuBuilder {
   companion object {
@@ -17,14 +20,23 @@ class WorkbenchMenuBuilder {
       useSystemMenuBarProperty().set(true)
     }
 
-    menuBar.menus.add(buildFileMenu())
+    menuBar.menus.add(buildFileMenu(scene))
     (scene.root as BorderPane).top = menuBar
   }
 
-  private fun buildFileMenu(): Menu {
+  private fun buildFileMenu(scene: Scene): Menu {
     val fileMenu = Menu(MENU_FILE)
-    val closeProjectMenuItem = MenuItem(MENU_ITEM_CLOSE_PROJECT)
+    val closeProjectMenuItem = buildProjectCloseMenuItem(scene)
     fileMenu.items.add(closeProjectMenuItem)
     return fileMenu
+  }
+
+  private fun buildProjectCloseMenuItem(scene: Scene): MenuItem {
+    return MenuItem(MENU_ITEM_CLOSE_PROJECT).apply {
+      setOnAction {
+        scene.closeWindow()
+        OpenRepoScene.launch(Stage())
+      }
+    }
   }
 }